no get_startup_event
